chore: Refactor main.py to include memory usage check and warning notification
