+++ containers/scraper/main.py
@@ -12,6 +12,7 @@ from selenium.webdriver.common.by import By
 from selenium.webdriver.chrome.options import Options
 from influxdb import InfluxDBClient
 import psycopg2
+from notifications import send_warning_notification, send_error_notification, send_error_notification_with_image
 
 # Constants
 WAIT_INTERVAL = 30
@@ -78,60 +79,6 @@ def write_to_influxdb(ticker, stock_price):
         }
     ]
     client.write_points(json_body)
-
-def send_warning_notification(message):
-    """Send error notification with a screenshot."""
-    with open(f"/output/{int(time.time())}.txt", "w") as f:
-        f.write(f"Warning: {message}")
-
-    response = requests.get("http://secret/Secret/AlertAccessToken")
-    access_token = response.json()["value"]
-    headers = {
-        "Content-Type": "application/x-www-form-urlencoded",
-        "Authorization": f"Bearer {access_token}"
-    }
-    payload = {
-        "message": f"Warning: {message}"
-    }
-    requests.post("https://notify-api.line.me/api/notify", headers=headers, params=payload)
-
-def send_error_notification(message):
-    """Send error notification with a screenshot."""
-    with open(f"/output/{int(time.time())}.txt", "w") as f:
-        f.write(f"Error: {message}")
-
-    response = requests.get("http://secret/Secret/ErrorAccessToken")
-    access_token = response.json()["value"]
-    headers = {
-        "Content-Type": "application/x-www-form-urlencoded",
-        "Authorization": f"Bearer {access_token}"
-    }
-    payload = {
-        "message": f"Error: {message}"
-    }
-    requests.post("https://notify-api.line.me/api/notify", headers=headers, params=payload)
-
-def send_error_notification_with_image(driver, e):
-    """Send error notification with a screenshot."""
-    with open(f"/output/{int(time.time())}.txt", "w") as f:
-        f.write(str(e))
-
-    screenshot_path = f"/output/{int(time.time())}.png"
-    driver.save_screenshot(screenshot_path)
-
-    response = requests.get("http://secret/Secret/ErrorAccessToken")
-    access_token = response.json()["value"]
-    headers = {
-        "Content-Type": "application/x-www-form-urlencoded",
-        "Authorization": f"Bearer {access_token}"
-    }
-    payload = {
-        "message": f"Error accessing element on URL: {driver.current_url}, Error: {e}"
-    }
-    files = {
-        "imageFile": open(screenshot_path, "rb")
-    }
-    requests.post("https://notify-api.line.me/api/notify", headers=headers, params=payload, files=files)
 
 def initialize_driver():
     """Initialize the Selenium WebDriver."""
