+++ containers/scraper/test.py
@@ -1,0 +1,4 @@
+from notifications import send_warning_notification, send_error_notification, send_error_notification_with_image
+
+send_warning_notification("This is a warning message.")
+send_error_notification("This is an error message.")
